host/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pyautogui==0.9.54
mss==10.0.0
python-dotenv==1.1.0
aiortc>=1.12.0
httpx>=0.28.0
websockets>=15.0
Pillow>=11.0.0
numpy>=2.0.0
deepgram-sdk>=3.0.0
openai>=1.0.0
pyperclip>=1.8.0
